perf: decide a cell's type once per column, not once per cell
Package 10 asked where Q11's 352 ms went, with a busy API and an idle
database. Into asking, 110,000 times, what type each cell was — a question
the column's PostgreSQL type OID had already answered before the first row
arrived.

`_plain()` ran once per cell in every format, up to four `isinstance` checks
deep. VOTable ran a second per-cell function on top of it and then
`xml.sax.saxutils.escape`: three `str.replace` calls, three allocations, on
text that is ObsCore identifiers and URLs and contains none of the three
characters it looks for.

A column's kind now names the Python type psycopg produces as well as the
wire type the writers emit, so the dispatch happens once per column:

- `float8`, `int*`, `bool` and the text family reach the writers untouched.
- `numeric` becomes its own kind (`decimal`): it arrives as a `Decimal` and
  is emitted as a VOTable double, and calling it `float64` was the reason
  the conversion had to be rediscovered per cell.
- An unrecognised OID — and a ColumnMeta nobody declared — is `opaque`, and
  keeps the fully dynamic path, which is what it needs.
- VOTable gets one text encoder per column and numbers skip escaping, since
  no decimal representation of a number contains an XML metacharacter.
- CSV rows go straight to `csv.writer`, which already renders None as an
  empty field and calls `str` in C.
- JSON encodes 500 rows per `dumps` rather than one.
- Arrow transposes with `zip(*batch)` instead of indexing every cell from
  Python.

Measured on the Q11 projection, 10,000 rows, best of 15 (µs/row):
arrow 8.01 → 0.93, parquet 8.77 → 1.78, json 15.44 → 5.12,
csv 17.05 → 9.19, tsv 17.25 → 9.13, votable 17.94 → 7.04.

The text output is byte-for-byte unchanged, which is the point:
`test_typed_and_opaque_columns_agree` asserts that declaring a type produces
the same bytes as not declaring it, over every kind, its boundary values, its
awkward representations and NULL. Parquet and Arrow changed in one respect,
for the better — a declared string column holding an exotic Python object
used to raise ArrowTypeError mid-stream and now serialises as text.

+++ tests/benchmarks/test_hot_paths.py
@@ -32,7 +32,7 @@ JOIN_QUERY = (
 COLUMNS = [
     ColumnMeta("source_id", kind="int64", ucd="meta.id"),
     ColumnMeta("name", kind="str"),
-    ColumnMeta("flux", kind="float64", unit="mJy"),
+    ColumnMeta("flux", kind="decimal", unit="mJy"),
     ColumnMeta("observed_at", kind="timestamp"),
     ColumnMeta("validated", kind="bool"),
 ]
